+++ src/types/extended.ts
@@ -6,14 +6,21 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-import { ViewContract } from './index.js'
+
+import { ViewContract, ViewRendererContract } from './main.js'
+
+declare module '@adonisjs/core/types' {
+  interface ContainerBindings {
+    view: ViewContract
+  }
+}
 
 declare module '@adonisjs/core/http' {
   /**
    * Decorate context
    */
   interface HttpContext {
-    view: ViewContract
+    view: ViewRendererContract
   }
 
   /**
