+++ src/types/main.ts
@@ -6,10 +6,20 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 import type { EdgeContract, EdgeRendererContract } from 'edge.js'
 
 export interface ViewContract extends EdgeContract {}
 export interface ViewRendererContract extends EdgeRendererContract {}
+
+/**
+ * Config accepted by the view provider
+ */
+export type ViewConfig = {
+  cache: {
+    enabled: boolean
+  }
+}
 
 export type {
   TagContract,
@@ -18,15 +28,3 @@ export type {
   TagTokenContract,
   TemplateConstructorContract,
 } from 'edge.js'
-
-/**
- * Config accepted by the view provider
- */
-export interface ViewConfig {
-  cache: {
-    enabled: boolean
-  }
-}
-
-export * from './container.js'
-export * from './context.js'
