--- types/container.ts
@@ -1,15 +1,0 @@
-/*
- * @adonisjs/view
- *
- * (c) AdonisJS
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-import { ViewContract } from './index.js'
-
-declare module '@adonisjs/core/types' {
-  interface ContainerBindings {
-    view: ViewContract
-  }
-}
